refactor(code): refactor file naming, export and imports
